+++ src/pdfnaut/objects/__init__.py
@@ -1,11 +1,12 @@
-from .actions import Action, ActionKind, Destination, DestinationKind, GoToAction
-from .annotations import Annotation, AnnotationFlags
+from .actions import Action, ActionKind, GoToAction, URIAction
+from .annotations import Annotation, AnnotationFlags, LinkAnnotation
 from .catalog import (
     PageLayout,
     PageMode,
     UserAccessPermissions,
     ViewerPreferences,
 )
+from .destinations import Destination, DestinationKind
 from .outlines import OutlineItem, OutlineItemFlags
 from .page import Page
 from .trailer import Info
@@ -17,10 +18,12 @@ __all__ = (
     "Destination",
     "DestinationKind",
     "GoToAction",
+    "URIAction",
     "PageLayout",
     "PageMode",
     "Page",
     "Annotation",
+    "LinkAnnotation",
     "AnnotationFlags",
     "Info",
     "UserAccessPermissions",